requirements.txt
appnope==0.1.0
attrs==19.1.0
backcall==0.1.0
bleach==3.1.0
certifi==2019.6.16
cycler==0.10.0
decorator==4.4.0
defusedxml==0.6.0
Django==2.2.5
entrypoints==0.3
imageio==2.5.0
jedi==0.15.1
Jinja2==2.10.1
joblib==0.13.2
jsonschema==3.0.2
kiwisolver==1.1.0
MarkupSafe==1.1.1
matplotlib==3.1.1
mistune==0.8.4
nbconvert==5.6.0
nbformat==4.4.0
networkx==2.3
notebook==6.0.0
numpy==1.17.0
opencv-python==4.1.0.25
pandocfilters==1.4.2
parso==0.5.1
pexpect==4.7.0
pickleshare==0.7.5
Pillow==6.1.0
prometheus-client==0.7.1
prompt-toolkit==2.0.9
ptyprocess==0.6.0
Pygments==2.4.2
pyparsing==2.4.2
pyrsistent==0.15.4
python-dateutil==2.8.0
pytz==2019.2
PyWavelets==1.0.3
pyzmq==18.1.0
qtconsole==4.5.3
scikit-image==0.15.0
scikit-learn==0.21.3
scikit-video==1.1.11
scipy==1.3.1
Send2Trash==1.5.0
six==1.12.0
sklearn==0.0
sqlparse==0.3.0
terminado==0.8.2
testpath==0.4.2
tornado==6.0.3
traitlets==4.3.2
wcwidth==0.1.7
webencodings==0.5.1
widgetsnbextension==3.5.1
